pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "docx-date-replacer"
version = "2.0.0"
description = "Утилита для пакетной замены даты в документах .docx"
requires-python = ">=3.10"
dependencies = [
    "python-docx>=1.2.0",
    "customtkinter>=5.2.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[project.scripts]
docx-date-replacer = "src.main:main"

[tool.setuptools.packages.find]
include = ["src*"]
